Moving logging state to utils, because it is no cli feature but more a utils feature
