+++ app/actions/install_actions/install_actions.go
@@ -3,12 +3,12 @@ package install_actions
 import (
 	"errors"
 	"fmt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/cli/secrets"
 	"k8s-management-go/app/constants"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/files"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 func ProgressNamespace(state models.StateData) (err error) {
