+++ app/actions/install_actions/install_nginx_ingress_controller.go
@@ -3,12 +3,12 @@ package install_actions
 import (
 	"errors"
 	"fmt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/constants"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/files"
 	"k8s-management-go/app/utils/helm"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"strconv"
 )
