+++ app/actions/uninstall_actions/uninstall_jenkins.go
@@ -2,11 +2,11 @@ package uninstall_actions
 
 import (
 	"fmt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/constants"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/helm"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 // uninstall Jenkins with Helm
