+++ app/cli/createproject/create_project_cloud_templates.go
@@ -3,11 +3,11 @@ package createproject
 import (
 	"fmt"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/constants"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/files"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 // project wizard dialog for cloud templates
