+++ app/cli/createproject/create_project_jobs_repo.go
@@ -3,9 +3,9 @@ package createproject
 import (
 	"errors"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"regexp"
 )
