+++ app/cli/createproject/create_project_processing.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 	"github.com/goware/prefixer"
 	"io/ioutil"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/constants"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/config"
 	"k8s-management-go/app/utils/files"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"os"
 	"strconv"
 	"strings"
