+++ app/cli/createproject/create_project_pvc.go
@@ -3,8 +3,8 @@ package createproject
 import (
 	"errors"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 func ProjectWizardAskForExistingPersistentVolumeClaim() (pvcName string, err error) {
