+++ app/cli/install/install_cli_workflow.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"k8s-management-go/app/actions/install_actions"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 /*
