+++ app/cli/secrets/apply_secrets.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"fmt"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"os"
 	"os/exec"
 )
