+++ app/cli/uninstall/uninstall_cli_workflow.go
@@ -3,8 +3,8 @@ package uninstall
 import (
 	"k8s-management-go/app/actions/uninstall_actions"
 	"k8s-management-go/app/cli/dialogs"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/models"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 // workflow for uninstall
