+++ app/gui/install/install_gui_workflow.go
@@ -4,10 +4,10 @@ import (
 	"fyne.io/fyne"
 	"fyne.io/fyne/dialog"
 	"k8s-management-go/app/actions/install_actions"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/gui/ui_elements"
 	"k8s-management-go/app/models"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"time"
 )
