+++ app/utils/encryption/bcrypt_jenkins_user_password.go
@@ -2,8 +2,8 @@ package encryption
 
 import (
 	"golang.org/x/crypto/bcrypt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 )
 
 func EncryptJenkinsUserPassword(plainPassword string) (hashedPassword string, err error) {
