+++ app/utils/encryption/gpg_encrypt.go
@@ -2,8 +2,8 @@ package encryption
 
 import (
 	"fmt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"os"
 	"os/exec"
 	"strings"
