+++ app/utils/kubectl/kubectl_executor.go
@@ -2,8 +2,8 @@ package kubectl
 
 import (
 	"fmt"
-	"k8s-management-go/app/cli/loggingstate"
 	"k8s-management-go/app/utils/logger"
+	"k8s-management-go/app/utils/loggingstate"
 	"os/exec"
 	"strings"
 )
